cairo_native_test

Function filter_test_case_compilation

pub(crate) fn filter_test_case_compilation(
    compiled: TestCompilation,
    compilation_filter: &[String],
) -> TestCompilation

Removes matching tests from TestCompilation. This not only includes the test cases, but also the function definition in the inner sierra program. This means that the matching tests are not even compiled.

DISCLAIMER: This is a hacky and temporary function, used to run corelib tests when not all libfuncs are implemented.